extract getPriceWithDiscount function and simplify order summary usage
