+++ support-frontend/assets/pages/paper-subscription-checkout/components/paperCheckoutForm.jsx
@@ -22,7 +22,7 @@ import Form, { FormSection, FormSectionHiddenUntilSelected } from 'components/ch
 import Layout, { Content } from 'components/subscriptionCheckouts/layout';
 import type { ErrorReason } from 'helpers/errorReasons';
 import { showPrice, type ProductPrices, type ProductPrice } from 'helpers/productPrice/productPrices';
-import { getProductPrice } from 'helpers/productPrice/paperProductPrices';
+import { getProductPrice, getPriceWithDiscount } from 'helpers/productPrice/paperProductPrices';
 import { HomeDelivery, Collection } from 'helpers/productPrice/fulfilmentOptions';
 import { formatMachineDate, formatUserDate } from 'helpers/dateConversions';
 import {
@@ -61,8 +61,7 @@ import type { IsoCurrency } from 'helpers/internationalisation/currency';
 import { withDeliveryFormIsValid } from 'helpers/subscriptionsForms/formValidation';
 import { setupSubscriptionPayPalPayment } from 'helpers/paymentIntegrations/payPalRecurringCheckout';
 import DirectDebitForm from 'components/directDebit/directDebitProgressiveDisclosure/directDebitForm';
-import { paperProductsWithDigital, paperProductsWithoutDigital, type ActivePaperProducts, type ProductOptions } from 'helpers/productPrice/productOptions';
-import { applyDiscount, getAppliedPromo } from 'helpers/productPrice/promotions';
+import { paperProductsWithDigital, paperProductsWithoutDigital, type ActivePaperProducts } from 'helpers/productPrice/productOptions';
 import { Paper } from 'helpers/subscriptions';
 import type { FulfilmentOptions } from 'helpers/productPrice/fulfilmentOptions';
 import DirectDebitPaymentTerms from 'components/subscriptionCheckouts/directDebit/directDebitPaymentTerms';
@@ -117,16 +116,6 @@ type PropTypes = {|
   fulfilmentOption: FulfilmentOptions,
 |};
 
-
-function getPricePlusDiscount(
-  productPrices: ProductPrices,
-  fulfilmentOption: FulfilmentOptions,
-  productOption: ProductOptions,
-) {
-  const basePrice = getProductPrice(productPrices, fulfilmentOption, productOption);
-  return applyDiscount(basePrice, getAppliedPromo(basePrice.promotions));
-}
-
 // ----- Map State/Props ----- //
 
 function mapStateToProps(state: WithDeliveryCheckoutState) {
@@ -142,7 +131,7 @@ function mapStateToProps(state: WithDeliveryCheckoutState) {
     csrf: state.page.csrf,
     currencyId: state.common.internationalisation.currencyId,
     payPalHasLoaded: state.page.checkout.payPalHasLoaded,
-    total: getPricePlusDiscount(
+    total: getPriceWithDiscount(
       state.page.checkout.productPrices,
       state.page.checkout.fulfilmentOption,
       state.page.checkout.productOption,
@@ -218,10 +207,10 @@ function PaperCheckoutForm(props: PropTypes) {
     // Price of the 'Plus' product that corresponds to the selected product option
     const plusPrice = includesDigiSub ?
       props.total :
-      getPricePlusDiscount(props.productPrices, props.fulfilmentOption, paperProductsWithDigital[props.productOption]);
+      getPriceWithDiscount(props.productPrices, props.fulfilmentOption, paperProductsWithDigital[props.productOption]);
     // Price of the standard paper-only product that corresponds to the selected product option
     const paperPrice = includesDigiSub ?
-      getPricePlusDiscount(
+      getPriceWithDiscount(
         props.productPrices,
         props.fulfilmentOption,
         paperProductsWithoutDigital[props.productOption],
